fix(runtime): guard spawn exit paths
