+++ crates/harness-server/src/task_runner/spawn/classification.rs
@@ -1,0 +1,62 @@
+/// Return `true` when a free-text prompt is complex enough to require a Plan phase
+/// before implementation.
+///
+/// Heuristic: prompt longer than 200 words OR contains 3 or more file-path-like
+/// tokens (sequences containing `/` or ending in a recognised source extension).
+pub fn prompt_requires_plan(prompt: &str) -> bool {
+    let word_count = prompt.split_whitespace().count();
+    if word_count > 200 {
+        return true;
+    }
+    let file_path_count = prompt
+        .split_whitespace()
+        .filter(|tok| {
+            // Exclude XML/HTML tags — `</foo>` contains '/' but is not a file path.
+            // System prompts wrap user data in `<external_data>...</external_data>`
+            // which would otherwise trigger false positives.
+            let is_xml_tag = tok.starts_with('<');
+            if is_xml_tag {
+                return false;
+            }
+            tok.contains('/') || {
+                let lower = tok.to_lowercase();
+                lower.ends_with(".rs")
+                    || lower.ends_with(".ts")
+                    || lower.ends_with(".tsx")
+                    || lower.ends_with(".go")
+                    || lower.ends_with(".py")
+                    || lower.ends_with(".toml")
+                    || lower.ends_with(".json")
+            }
+        })
+        .count();
+    file_path_count >= 3
+}
+
+/// Extract `(owner, repo, number)` from a GitHub PR URL.
+///
+/// Expects format: `https://github.com/{owner}/{repo}/pull/{number}[#...]`
+pub(crate) fn parse_pr_url(pr_url: &str) -> Option<(String, String, u64)> {
+    // Strip fragment (e.g. #discussion_r...)
+    let url = pr_url.split('#').next().unwrap_or(pr_url);
+    let parts: Vec<&str> = url.trim_end_matches('/').split('/').collect();
+    // Expected: ["https:", "", "github.com", owner, repo, "pull", number]
+    let pull_idx = parts.iter().rposition(|&p| p == "pull")?;
+    if pull_idx + 1 >= parts.len() || pull_idx < 2 {
+        return None;
+    }
+    let number: u64 = parts[pull_idx + 1].parse().ok()?;
+    let repo = parts[pull_idx - 1].to_string();
+    let owner = parts[pull_idx - 2].to_string();
+    Some((owner, repo, number))
+}
+
+/// Convert a user-facing timeout value to a Duration.
+/// `0` means "no timeout" and maps to ~277 hours (effectively unlimited).
+pub fn effective_turn_timeout(secs: u64) -> tokio::time::Duration {
+    if secs == 0 {
+        tokio::time::Duration::from_secs(999_999)
+    } else {
+        tokio::time::Duration::from_secs(secs)
+    }
+}
