+++ crates/harness-server/src/task_runner/spawn/project.rs
@@ -1,0 +1,59 @@
+use std::path::{Path, PathBuf};
+
+use super::super::request::detect_main_worktree;
+
+fn describe_detect_main_worktree_join_error(join_err: &tokio::task::JoinError) -> String {
+    if join_err.is_panic() {
+        format!("detect_main_worktree panicked: {join_err}")
+    } else if join_err.is_cancelled() {
+        format!("detect_main_worktree was cancelled: {join_err}")
+    } else {
+        format!("detect_main_worktree failed: {join_err}")
+    }
+}
+
+pub(super) async fn resolve_project_root_with(
+    requested_project: Option<PathBuf>,
+    detect_worktree: impl FnOnce() -> PathBuf + Send + 'static,
+) -> anyhow::Result<PathBuf> {
+    match requested_project {
+        Some(project) => Ok(project),
+        None => tokio::task::spawn_blocking(detect_worktree)
+            .await
+            .map_err(|join_err| {
+                let reason = describe_detect_main_worktree_join_error(&join_err);
+                tracing::error!("{reason}");
+                anyhow::anyhow!("{reason}")
+            }),
+    }
+}
+
+pub(super) fn validate_spawn_project_root(
+    raw_project: &Path,
+    home_dir: &Path,
+    allowed_project_roots: &[PathBuf],
+) -> Result<PathBuf, String> {
+    if allowed_project_roots.is_empty() {
+        return crate::handlers::validate_project_root(raw_project, home_dir);
+    }
+
+    let canonical = raw_project
+        .canonicalize()
+        .map_err(|e| format!("invalid project root '{}': {e}", raw_project.display()))?;
+    if !canonical.is_dir() {
+        return Err(format!(
+            "project root is not a directory: {}",
+            canonical.display()
+        ));
+    }
+    crate::project_registry::check_allowed_roots(&canonical, allowed_project_roots)?;
+    Ok(canonical)
+}
+
+pub async fn resolve_canonical_project(project: Option<PathBuf>) -> anyhow::Result<PathBuf> {
+    let raw = resolve_project_root_with(project, detect_main_worktree).await?;
+    // Best-effort canonicalize: if the path doesn't exist yet (e.g. in tests
+    // using a path that will be created later) fall back to the raw path so
+    // we at least get a consistent string key.
+    Ok(raw.canonicalize().unwrap_or(raw))
+}
